refactor(stella-tui): split deck.rs's pure event classifiers into deck/classify.rs
`deck.rs` sat at 1492 of the 1500-line guard, and #2007's parked-wait clock
needs about fourteen lines in it. The sanctioned move is to extract a coherent
cluster rather than raise a ceiling, exactly as `prompt_queue.rs` was split out
of this same file for the same reason.

The cluster is the file's "Event → derived attributes" section: four pure
functions over `&AgentEvent` that hold no state — `event_intensity`,
`status_from_event`, `trace_of` and `snip`. They were private, called only from
within `deck.rs`, so the cut needs no visibility change beyond the `pub(super)`
a child module requires.

A pure move: the function bodies are byte-identical, `deck.rs` re-imports the
four names so every call site is unchanged, and no test changed. deck.rs drops
to 1207 lines.

Refs #2007
